add unit tests for helper functions
